Puerto rico democrat results
Not sure how much I trust these numbers. only 51% is reporting according
to that source and that number is way different than the one reported by
the nytimes on their delegate counter page. Which has a much lower
number of delegates for each.

AP says that they took the day off:

	http://hosted.ap.org/dynamic/stories/U/US_2016_ELECTION_PUERTO_RICO?SITE=CASON&SECTION=STATE&TEMPLATE=DEFAULT&CTIME=2016-06-06-11-58-24

So I guess I have to wait a bit and will have to update these numbers.

nytimes says 36 Clinton and 24 Sanders for delegates, but doesn't have a
results page for it for some reason.